ISSUE-119: Abstract integration test logic into base class
The unit tests have a base class in which there is a `runCommand`
method. I have created similar logic for the integration tests.

I have migrated `FirstTest.php` to `HelloWorldTest.php` as that was
the command it was testing.

Assuming we are using the pestle package in the installs environment,
I have allowed the user to name there package 3 names and added a
basic fallback structure.